feature: added trace settings saving
