+++ example/STMViewer_test/STMViewer_project/STMViewer_test.cfg
@@ -22,6 +22,10 @@ max_viewport_points_percent = 10
 trigger_channel = -1
 trigger_level = 0.000000
 timeout = 2
+probe_type = 1
+target_name = 
+probe_speed_khz = 10000
+probe_sn = 506003225
 
 [var0]
 name = test.a
